Show full month names for all event dates
Update date formatting in multiple components and utilities to display the full month name (e.g., "March" instead of "Mar") instead of the abbreviated version.

+++ client/src/components/sports/SportEventCard.tsx
@@ -162,7 +162,7 @@ const SportEventCard: React.FC<SportEventCardProps> = ({ event, sportId }) => {
       return `Today, ${date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}`;
     }
     
-    return date.toLocaleDateString([], { weekday: 'short', month: 'short', day: 'numeric' }) + 
+    return date.toLocaleDateString([], { weekday: 'short', month: 'long', day: 'numeric' }) + 
       ', ' + date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
   };
