+++ client/src/pages/sport-page.tsx
@@ -55,7 +55,7 @@ const SportEventCard = ({ event, sportId }: { event: Event, sportId: number }) =
       return `Today, ${date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' })}`;
     }
     
-    return date.toLocaleDateString([], { weekday: 'short', month: 'short', day: 'numeric' }) + 
+    return date.toLocaleDateString([], { weekday: 'short', month: 'long', day: 'numeric' }) + 
       ', ' + date.toLocaleTimeString([], { hour: '2-digit', minute: '2-digit' });
   };
